refactor(systems): split MapSystem types and generation helpers
Extract MapSystem.types.ts (level/chunk/progression contracts) and
MapSystem.generation.ts (path/spawn utilities, defaults) from the
monolithic MapSystem.ts. MapSystem.ts keeps managers and re-exports
the public API unchanged so TrackManager/registry imports stay stable.

Also document registry.contract.ts in TESTING.md — the compile-time +
runtime registry drift guard landed in #319; this completes the optional
MapSystem split from issue #313.
